backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
python-dotenv==1.1.1
pydantic==2.11.9
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
